# swagger/swagger.yaml
swagger: "2.0"
info:
  description: "This is a sample swagger doc detailing the supplier POST route for learning"
  version: "0.0.1"
  title: "Distributable Swagger Doc"
  contact:
    email: "im.solomon@Outlook.com"
securityDefinitions:
  Bearer:
    type: apiKey
    name: Authorization
    in: header
    description: |
     For accessing the API a valid JWT token must be passed in all the queries in
     the 'Authorization' header.
  
  
     A valid JWT token is generated by the API and retourned as answer of a call
     to the route /login giving a valid user & password.
  
  
     The following syntax must be used in the 'Authorization' header :
  
         Bearer xxxxxx.yyyyyyy.zzzzzz
host: "localhost:8080"
basePath: /api/v1
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
...
paths:
  /countries:    
    x-swagger-router-controller: countries
    get:
      tags:
      - "GeneralData"
      summary: "Get a list of countries"
      operationId: getCountries
      security:
        - Bearer: []
      x-security-scopes:
        - user
      produces:
      - "application/json"
      parameters:
      - in: "query"
        name: "name"
        description: "Name of the country"
        type: "string"
      - in: "query"
        name: "field"
        description: "Limit the fields to be returned"
        type: "array"
        uniqueItems: true
        items:
          type: string
        default:
          - '*'
      - in: "query"
        name: "sort"
        description: "Field on which to sort the returned countries"
        type: "string"
      - in: "query"
        name: "order"
        description: "Order by which to order the sorted countries"
        type: "string"
        default: "ASC"
        enum:
          - "ASC"
          - "DESC"
      # - in: "query"
      #   name: "limit"
      #   description: "Limit the number of returned result"
      #   type: "integer"
      #   format: "int32"
      #   default: 50
      #   minimum: 1
      # - in: "query"
        # name: "offset"
        # description: "Offset the number of returned result from an index of 0"
        # type: "integer"
        # format: "int64"
        # default: 0
        # minimum: 0
      responses:
        200:
          description: "Successfully Retrieved Countries"
          schema:
            type: "object"
            properties:
              results:
                type: "array"
                items:
                  $ref: "#/definitions/Country"
              message: 
                type: "string"
                description: "If empty result, it shall state 'No results were found'. If results found, it shall state 'Found X result(s)"
              total:
                type: "integer"
                description: "The actual number of suppliers found"
              # offset:
              #   type: "integer"
              #   description: "The returned return offset from an index of 0"
              # limit:
              #   type: "integer"
              #   description: "Limits the number of results returned"
        400:
          description: "Invalid Input"
          schema:
            type: "object"
            properties:
              status:
                type: "integer"
                description: "The http status code"
              code:
                type: "integer"
                description: "The error code"
              message:
                type: "string"
                description: "The message associated to the error code"
        403:
          description: "Access Denied"
          schema:
            type: "object"
            properties:
              status:
                type: "integer"
                description: "The http status code"
              code:
                type: "integer"
                description: "The error code"
              message:
                type: "string"
                description: "The message associated to the error code"
...
definitions:
...
  Country:
    type: "object"
    description: "The model for a country as returned by the API"
    properties:
      code:
        type: "string"        
        description: "The country code in the system"
      name:
        type: "string"
        description: "The name of the country"
...